refactor classes
