+++ Laboratory1/app/src/main/AndroidManifest.xml
@@ -11,14 +11,14 @@
         android:roundIcon="@mipmap/airplane_ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.Laboratory1">
-        <activity android:name=".GameActivity"></activity>
-        <activity android:name=".SecondActivity" />
-        <activity android:name=".SixthActivity" />
+        <activity android:name=".presenter.game.GameActivity"></activity>
+        <activity android:name=".presenter.game.SecondActivity" />
+        <activity android:name=".presenter.users.SixthActivity" />
         <activity
-            android:name=".FourthActivity"
-            android:parentActivityName=".ThirdActivity" />
-        <activity android:name=".FifthActivity" />
-        <activity android:name=".ThirdActivity">
+            android:name=".presenter.randomise.FourthActivity"
+            android:parentActivityName=".presenter.ThirdActivity" />
+        <activity android:name=".presenter.posts.FifthActivity" />
+        <activity android:name=".presenter.ThirdActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
