--- Laboratory1/app/src/main/java/com/example/laboratory1/Position.kt
@@ -1,3 +1,0 @@
-package com.example.laboratory1
-
-data class Position(val row: Int, val column: Int)
