+++ Laboratory1/app/src/main/java/com/example/laboratory1/api/ApiHelper.kt
@@ -1,7 +1,6 @@
-package com.example.laboratory1.data.api
+package com.example.laboratory1.api
 
 class ApiHelper(private val apiService: ApiService) {
 
     suspend fun getUsers() = apiService.getUsers()
-    suspend fun getPosts() = apiService.getPosts()
 }
