+++ Laboratory1/app/src/main/java/com/example/laboratory1/api/ApiService.kt
@@ -1,0 +1,10 @@
+package com.example.laboratory1.api
+
+import com.example.laboratory1.api.model.User
+import retrofit2.http.GET
+
+interface ApiService {
+
+    @GET("users")
+    suspend fun getUsers(): List<User>
+}
