+++ Laboratory1/app/src/main/java/com/example/laboratory1/api/status/Resource.kt
@@ -1,4 +1,4 @@
-package com.example.laboratory1.utils
+package com.example.laboratory1.api.status
 
 data class Resource<out T>(val status: Status, val data: T?, val message: String?) {
     companion object {
