+++ Laboratory1/app/src/main/java/com/example/laboratory1/api/status/Status.kt
@@ -1,4 +1,4 @@
-package com.example.laboratory1.utils
+package com.example.laboratory1.api.status
 
 enum class Status {
     SUCCESS,
