--- Laboratory1/app/src/main/java/com/example/laboratory1/data/model/Posts.kt
@@ -1,7 +1,0 @@
-package com.example.laboratory1.data.model
-
-data class Posts(
-        val title: String,
-        val id: String,
-        val body: String
-)
