--- Laboratory1/app/src/main/java/com/example/laboratory1/data/repository/MainRepository.kt
@@ -1,9 +1,0 @@
-package com.example.laboratory1.data.repository
-
-import com.example.laboratory1.data.api.ApiHelper
-
-class MainRepository(private val apiHelper: ApiHelper) {
-
-    suspend fun getUsers() = apiHelper.getUsers()
-    suspend fun getPosts() = apiHelper.getPosts()
-}
