+++ Laboratory1/app/src/main/java/com/example/laboratory1/presenter/game/model/Position.kt
@@ -1,0 +1,5 @@
+package com.example.laboratory1.presenter.game.model
+
+data class Position(
+        val row: Int,
+        val column: Int)
