+++ Laboratory1/app/src/main/java/com/example/laboratory1/presenter/randomise/FourthActivity.kt
@@ -1,9 +1,14 @@
-package com.example.laboratory1
+package com.example.laboratory1.presenter.randomise
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.example.laboratory1.R
 import com.example.laboratory1.databinding.ActivityFourthBinding
+import com.example.laboratory1.presenter.ThirdActivity
+import com.example.laboratory1.presenter.game.SecondActivity
+import com.example.laboratory1.presenter.posts.FifthActivity
+import com.example.laboratory1.presenter.users.SixthActivity
 import com.google.android.material.bottomnavigation.BottomNavigationView
 
 class FourthActivity : AppCompatActivity() {
@@ -46,24 +51,19 @@ class FourthActivity : AppCompatActivity() {
             when (item.itemId) {
                 R.id.my_profile -> {
                     val intent = Intent(this, ThirdActivity::class.java)
-                    startActivity(intent)
-                }
+                    startActivity(intent) }
                 R.id.posts -> {
                     val intent = Intent(this, FifthActivity::class.java)
-                    startActivity(intent)
-                }
+                    startActivity(intent) }
                 R.id.randomise -> {
                     val intent = Intent(this, FourthActivity::class.java)
-                    startActivity(intent)
-                }
+                    startActivity(intent) }
                 R.id.users -> {
                     val intent = Intent(this, SixthActivity::class.java)
-                    startActivity(intent)
-                }
+                    startActivity(intent) }
                 R.id.games -> {
                     val intent = Intent(this, SecondActivity::class.java)
-                    startActivity(intent)
-                }
+                    startActivity(intent) }
             }
             true
         }
