+++ Laboratory1/app/src/main/java/com/example/laboratory1/presenter/users/SixthActivity.kt
@@ -1,4 +1,4 @@
-package com.example.laboratory1
+package com.example.laboratory1.presenter.users
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
@@ -9,26 +9,25 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProviders
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import com.example.laboratory1.data.api.ApiHelper
-import com.example.laboratory1.data.api.RetrofitBuilder
-import com.example.laboratory1.data.model.Posts
-import com.example.laboratory1.data.model.User
-import com.example.laboratory1.ui.base.ViewModelFactory
-import com.example.laboratory1.ui.base.ViewModelFactory1
-import com.example.laboratory1.ui.main.adapter.MainAdapter
-import com.example.laboratory1.ui.main.adapter.PostsAdapter
-import com.example.laboratory1.ui.main.viewmodel.MainViewModel
-import com.example.laboratory1.ui.main.viewmodel.PostsViewModel
-import com.example.laboratory1.utils.Status
+import com.example.laboratory1.R
+import com.example.laboratory1.api.ApiHelper
+import com.example.laboratory1.api.RetrofitBuilder
+import com.example.laboratory1.api.model.User
+import com.example.laboratory1.api.ViewModelFactory
+import com.example.laboratory1.api.MainAdapter
+import com.example.laboratory1.api.view.MainViewModel
+import com.example.laboratory1.api.status.Status
+import com.example.laboratory1.presenter.posts.FifthActivity
+import com.example.laboratory1.presenter.randomise.FourthActivity
+import com.example.laboratory1.presenter.game.SecondActivity
+import com.example.laboratory1.presenter.ThirdActivity
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import kotlinx.android.synthetic.main.activity_sixth.*
 
 class SixthActivity : AppCompatActivity() {
 
     private lateinit var viewModel: MainViewModel
-    private lateinit var viewModel1: PostsViewModel
     private lateinit var adapter: MainAdapter
-    private lateinit var adapter1: PostsAdapter
 
     lateinit var bottomNavigation : BottomNavigationView
 
@@ -37,11 +36,8 @@ class SixthActivity : AppCompatActivity() {
         setContentView(R.layout.activity_sixth)
 
         setupViewModel()
-        setupViewModel1()
         setupUI()
-        setupUI1()
         setupObservers()
-        setupObservers1()
 
         bottomNavigation = findViewById(R.id.bottom_navigation)
 
@@ -82,12 +78,6 @@ class SixthActivity : AppCompatActivity() {
         ).get(MainViewModel::class.java)
     }
 
-    private fun setupViewModel1() {
-        viewModel1 = ViewModelProviders.of(
-                this,
-                ViewModelFactory1(ApiHelper(RetrofitBuilder.apiService))
-        ).get(PostsViewModel::class.java)
-    }
 
     private fun setupUI() {
         recyclerView.layoutManager = LinearLayoutManager(this)
@@ -97,18 +87,6 @@ class SixthActivity : AppCompatActivity() {
                 recyclerView.context,
                 (recyclerView.layoutManager as LinearLayoutManager).orientation
             )
-        )
-        recyclerView.adapter = adapter
-    }
-
-    private fun setupUI1() {
-        recyclerView.layoutManager = LinearLayoutManager(this)
-        adapter1 = PostsAdapter(arrayListOf())
-        recyclerView.addItemDecoration(
-                DividerItemDecoration(
-                        recyclerView.context,
-                        (recyclerView.layoutManager as LinearLayoutManager).orientation
-                )
         )
         recyclerView.adapter = adapter
     }
@@ -136,39 +114,9 @@ class SixthActivity : AppCompatActivity() {
         })
     }
 
-    private fun setupObservers1() {
-        viewModel1.getPosts().observe(this, Observer {
-            it?.let { resource ->
-                when (resource.status) {
-                    Status.SUCCESS -> {
-                        recyclerView.visibility = View.VISIBLE
-                        progressBar.visibility = View.GONE
-                        resource.data?.let { posts -> retrieveList1(posts) }
-                    }
-                    Status.ERROR -> {
-                        recyclerView.visibility = View.VISIBLE
-                        progressBar.visibility = View.GONE
-                        Toast.makeText(this, it.message, Toast.LENGTH_LONG).show()
-                    }
-                    Status.LOADING -> {
-                        progressBar.visibility = View.VISIBLE
-                        recyclerView.visibility = View.GONE
-                    }
-                }
-            }
-        })
-    }
-
     private fun retrieveList(users: List<User>) {
         adapter.apply {
             addUsers(users)
-            notifyDataSetChanged()
-        }
-    }
-
-    private fun retrieveList1(posts: List<Posts>) {
-        adapter1.apply {
-            addPosts(posts)
             notifyDataSetChanged()
         }
     }
