+++ Laboratory1/app/src/main/java/com/example/laboratory1/recycleView/RecycleAdapter.kt
@@ -3,12 +3,8 @@ package com.example.laboratory1.recycleView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageView
-import android.widget.TextView
 import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
-import com.example.laboratory1.R
 import com.example.laboratory1.databinding.ItemLayoutBinding
 
 class RecycleAdapter(private var titles: List<String>,
