+++ Laboratory1/app/src/main/java/com/example/laboratory1/utils/AppConstants.kt
@@ -1,0 +1,7 @@
+package com.example.laboratory1.utils
+
+object AppConstants {
+
+    internal const val BASE_URL = "https://jsonplaceholder.typicode.com/"
+
+}
