+++ Laboratory1/app/src/main/res/layout/activity_fourth.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FourthActivity">
+    tools:context=".presenter.randomise.FourthActivity">
 
     <ImageView
         android:id="@+id/imageView2"
