+++ Laboratory1/app/src/main/res/layout/activity_second.xml
@@ -5,7 +5,7 @@ xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".SecondActivity">
+tools:context=".presenter.game.SecondActivity">
 
     <ImageView
         android:id="@+id/imageView6"
